requirements.txt
torch
torchvision
torchmetrics
numpy
matplotlib
scipy
opencv-python
pillow
scikit-learn
vit-pytorch